+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/test/java/org/apache/hadoop/yarn/server/resourcemanager/webapp/TestRMWebServicesCapacitySched.java
@@ -18,39 +18,19 @@
 
 package org.apache.hadoop.yarn.server.resourcemanager.webapp;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import javax.ws.rs.core.MediaType;
+
 import com.google.inject.Guice;
 import com.google.inject.servlet.ServletModule;
 import com.sun.jersey.api.client.ClientResponse;
 import com.sun.jersey.api.client.WebResource;
 import com.sun.jersey.guice.spi.container.servlet.GuiceContainer;
-import com.sun.jersey.test.framework.WebAppDescriptor;
-
-import java.io.BufferedReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.StringReader;
-import java.io.StringWriter;
-import java.net.URISyntaxException;
-import java.util.Objects;
-import java.util.stream.Collectors;
-
-import javax.ws.rs.core.MediaType;
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.transform.OutputKeys;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerException;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
+import org.codehaus.jettison.json.JSONObject;
+import org.junit.Before;
+import org.junit.Test;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.http.JettyUtils;
-import org.apache.hadoop.util.XMLUtils;
 import org.apache.hadoop.yarn.api.records.Resource;
 import org.apache.hadoop.yarn.conf.YarnConfiguration;
 import org.apache.hadoop.yarn.server.resourcemanager.MockRM;
@@ -64,19 +44,8 @@ import org.apache.hadoop.yarn.util.resource.Resources;
 import org.apache.hadoop.yarn.webapp.GenericExceptionHandler;
 import org.apache.hadoop.yarn.webapp.GuiceServletConfig;
 import org.apache.hadoop.yarn.webapp.JerseyTestBase;
-import org.codehaus.jettison.json.JSONException;
-import org.codehaus.jettison.json.JSONObject;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.w3c.dom.Document;
-import org.xml.sax.InputSource;
 
-import static org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration.AUTO_CREATED_LEAF_QUEUE_TEMPLATE_PREFIX;
-import static org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration.CAPACITY;
-import static org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration.DOT;
-import static org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration.MAX_PARALLEL_APPLICATIONS;
-import static org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration.PREFIX;
+import static org.apache.hadoop.yarn.server.resourcemanager.webapp.TestWebServiceUtil.*;
 import static org.junit.Assert.assertEquals;
 
 public class TestRMWebServicesCapacitySched extends JerseyTestBase {
@@ -114,7 +83,7 @@ public class TestRMWebServicesCapacitySched extends JerseyTestBase {
         Guice.createInjector(new WebServletModule(rm)));
   }
 
-  public CapacitySchedulerConfiguration setupQueueConfiguration(
+  public static CapacitySchedulerConfiguration setupQueueConfiguration(
       CapacitySchedulerConfiguration config) {
     config.set("yarn.scheduler.capacity.root.queues", "a, b, c");
     config.set("yarn.scheduler.capacity.root.a.queues", "a1, a2");
@@ -274,116 +243,9 @@ public class TestRMWebServicesCapacitySched extends JerseyTestBase {
     assertEquals("<memory:10, vCores:1>", res.toString());
   }
 
-  public static void assertXmlType(ClientResponse response) {
-    assertEquals(MediaType.APPLICATION_XML_TYPE + "; " + JettyUtils.UTF_8,
-        response.getType().toString());
-  }
-
-  public static void assertXmlResponse(ClientResponse response,
-                                       String expectedResourceFilename) throws
-      Exception {
-    assertXmlType(response);
-    Document document = loadDocument(response.getEntity(String.class));
-    String actual = serializeDocument(document).trim();
-    updateTestDataAutomatically(expectedResourceFilename, actual);
-    assertEquals(getResourceAsString(expectedResourceFilename), actual);
-  }
-
-  public static String serializeDocument(Document document) throws TransformerException {
-    DOMSource domSource = new DOMSource(document);
-    StringWriter writer = new StringWriter();
-    StreamResult result = new StreamResult(writer);
-    TransformerFactory tf = XMLUtils.newSecureTransformerFactory();
-    Transformer transformer = tf.newTransformer();
-    transformer.setOutputProperty(OutputKeys.INDENT, "yes");
-    transformer.setOutputProperty("{http://xml.apache.org/xslt}indent-amount", "2");
-    transformer.transform(domSource, result);
-    return writer.toString();
-  }
-
-  public static Document loadDocument(String xml) throws Exception {
-    DocumentBuilderFactory factory = XMLUtils.newSecureDocumentBuilderFactory();
-    DocumentBuilder builder = factory.newDocumentBuilder();
-    InputSource is = new InputSource(new StringReader(xml));
-    return builder.parse(is);
-  }
-
-  public static void assertJsonResponse(ClientResponse response,
-                                        String expectedResourceFilename) throws
-      JSONException, IOException {
-    assertJsonType(response);
-    JSONObject json = response.getEntity(JSONObject.class);
-    String actual = json.toString(2);
-    updateTestDataAutomatically(expectedResourceFilename, actual);
-    assertEquals(
-        prettyPrintJson(getResourceAsString(expectedResourceFilename)),
-        prettyPrintJson(actual));
-  }
-
-  private static String prettyPrintJson(String in) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper();
-    return objectMapper
-        .writerWithDefaultPrettyPrinter()
-        .writeValueAsString(objectMapper.readTree(in));
-  }
-
-  public static void assertJsonType(ClientResponse response) {
-    assertEquals(MediaType.APPLICATION_JSON_TYPE + "; " + JettyUtils.UTF_8,
-        response.getType().toString());
-  }
-
-  public static InputStream getResourceAsStream(String configFilename) {
-    ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-    return classLoader.getResourceAsStream(configFilename);
-  }
-
-  public static String getResourceAsString(String configFilename) throws IOException {
-    try (InputStream is = getResourceAsStream(configFilename)) {
-      if (is == null) {
-        return null;
-      }
-      try (InputStreamReader isr = new InputStreamReader(is);
-           BufferedReader reader = new BufferedReader(isr)) {
-        return reader.lines().collect(Collectors.joining(System.lineSeparator()));
-      }
-    }
-  }
-
-  public static void updateTestDataAutomatically(String configFilename, String actualContent) {
-    /*
-     Set UPDATE_TESTDATA=1 environment variable for auto update the expected data
-     or uncomment this return statement.
-
-     It's safe in a way that, this updates the source directory so the test will still fail,
-     because the target directory is untouched.
-     */
-    if (System.getenv("UPDATE_TESTDATA") == null) {
-      return;
-    }
-
-    ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-    try {
-      String resource = Objects.requireNonNull(
-          Objects.requireNonNull(classLoader.getResource(configFilename)).toURI().getPath())
-          .replaceAll("/target/test-classes/", "/src/test/resources/");
-      try (FileWriter writer = new FileWriter(resource, false)) {
-        writer.write(actualContent);
-      }
-    } catch (URISyntaxException | IOException e) {
-      e.printStackTrace();
-      Assert.fail("overwrite should not fail " + e.getMessage());
-    }
-  }
-
-  public static WebAppDescriptor createWebAppDescriptor() {
-    return new WebAppDescriptor.Builder(
-        TestRMWebServicesCapacitySched.class.getPackage().getName())
-        .contextListenerClass(GuiceServletConfig.class)
-        .filterClass(com.google.inject.servlet.GuiceFilter.class)
-        .contextPath("jersey-guice-filter").servletPath("/").build();
-  }
 
   public static MockRM createMockRM(CapacitySchedulerConfiguration csConf) {
+    setupQueueConfiguration(csConf);
     YarnConfiguration conf = new YarnConfiguration(csConf);
     conf.setClass(YarnConfiguration.RM_SCHEDULER, CapacityScheduler.class,
         ResourceScheduler.class);
